pyproject.toml
[project]
name = "memoant"
version = "0.1.0"
description = "Record meetings, transcribe voice memos, structure everything into Oracle DB + Obsidian notes"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "watchdog>=4.0",
    "mlx-whisper>=0.4",
    "silero-vad>=5.1",
    "pyannote-audio>=3.3",
    "torch>=2.2",
    "torchaudio>=2.2",
    "python-dotenv>=1.0",
    "click>=8.1",
    "tomli>=2.0",
]

[project.scripts]
memoant = "memoant.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memoant"]
